+++ src/ladb_opencutlist/js/plugins/jquery.ladb.tab-forum.js
@@ -196,6 +196,10 @@
         // Bind buttons
         $('#ladb_forum_redirection_continue', $modal).on('click', function () {
             callback();
+
+            // hide modal
+            $modal.modal('hide');
+
             return false;
         });