+++ src/ladb_opencutlist/ruby/model/cutlist/cutlist.rb
@@ -6,10 +6,10 @@ module Ladb::OpenCutList
 
     include HashableHelper
 
-    attr_accessor :selection_only, :length_unit, :currency_symbol, :mass_unit_strippedname, :dir, :filename, :model_name, :page_label, :max_number, :instance_count, :ignored_instance_count, :solid_wood_material_count, :sheet_good_material_count, :dimensional_material_count, :edge_material_count, :hardware_material_count
+    attr_accessor :selection_only, :length_unit, :currency_symbol, :mass_unit_strippedname, :dir, :filename, :model_name, :model_description, :page_label, :max_number, :instance_count, :ignored_instance_count, :solid_wood_material_count, :sheet_good_material_count, :dimensional_material_count, :edge_material_count, :hardware_material_count
     attr_reader :errors, :warnings, :tips, :used_tags, :material_usages, :groups
 
-    def initialize(selection_only, length_unit, mass_unit_strippedname, currency_symbol, dir, filename, model_name, page_label, instance_count)
+    def initialize(selection_only, length_unit, mass_unit_strippedname, currency_symbol, dir, filename, model_name, model_description, page_label, instance_count)
       @_obsolete = false
       @_observers = []
 
@@ -23,6 +23,7 @@ module Ladb::OpenCutList
       @dir = dir
       @filename = filename
       @model_name = model_name
+      @model_description = model_description
       @page_label = page_label
       @instance_count = instance_count
       @ignored_instance_count = 0