+++ src/ladb_opencutlist/ruby/model/report/report_entry.rb
@@ -209,6 +209,8 @@ module Ladb::OpenCutList
 
       @id = _def.cutlist_part.id
       @name = _def.cutlist_part.name
+      @unused_instance_count = _def.cutlist_part.unused_instance_count
+      @instance_count_by_part = _def.cutlist_part.instance_count_by_part
       @flipped = _def.cutlist_part.flipped
       @count = _def.cutlist_part.count