+++ src/ladb_opencutlist/ruby/worker/cutlist/cutlist_generate_worker.rb
@@ -84,10 +84,11 @@ module Ladb::OpenCutList
       currency_symbol = PriceUtils.instance.get_symbol
       dir, filename = File.split(model && !model.path.empty? ? model.path : Plugin.instance.get_i18n_string('default.empty_filename'))
       model_name = model ? model.name : ''
+      model_description = model ? model.description : ''
       page_label = model && model.pages && model.pages.selected_page ? model.pages.selected_page.label : ''
 
       # Create cut list
-      cutlist = Cutlist.new(selection_only, length_unit, mass_unit_strippedname, currency_symbol, dir, filename, model_name, page_label, @instance_infos_cache.length)
+      cutlist = Cutlist.new(selection_only, length_unit, mass_unit_strippedname, currency_symbol, dir, filename, model_name, model_description, page_label, @instance_infos_cache.length)
 
       # Errors & tips
       if @instance_infos_cache.length == 0