+++ CHANGELOG.md
@@ -10,6 +10,7 @@ in 2.x, 1.x and 0.x versions.
   * Fixed forum conversation tag filter
   * Fixed bug in preset names with quote
   * Added total cut length in cutting diagram summary
+  * Added model description in printable headers if it's defined
   
 * 2.1.0 (2021-05-28)